Initialization of default log_level moved to library
Previously application needed to initialize which puts an unnecessary
dependency onto something which is not obvious.

If there are future and currently unknown advantages to this
being initialized by the application, corresponding code in library should
be  conditionally build using build variables. Default should still be the
behaviour as now. This is not very likely as DEF_LOG_LEVEL already provides
build/configure flexibility.

Exception should be if library is compiled as shared library and different
behaviour is expected from different apps/libs (i.e. if reused by the
project it originates from libbladerf which may have applications expecting
a certain behaviour).